moved vcr to make_vcr
