+++ tests/poll/test_views.py
@@ -15,7 +15,7 @@ from bulbs.utils.test import (
     mock_vault,
     BaseAPITestCase,
 )
-from bulbs.utils.test.vcr import make_vcr
+from bulbs.utils.test.make_vcr import make_vcr
 from .common import SECRETS
 
 vcr = make_vcr(__file__)  # Define vcr file path
